requirements.txt
fastapi
uvicorn
# torch
# torchvision
pillow
onnxruntime
numpy